+++ templates/Admin/Users/index.php
@@ -11,7 +11,10 @@ declare(strict_types=1);
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ *
+ * @var \MeCms\View\View\AdminView $this
  */
+
 $this->extend('MeCms./Admin/common/index');
 $this->assign('title', I18N_USERS);
 $this->append('actions', $this->Html->button(
@@ -99,7 +102,7 @@ echo $this->Form->end();
                     ];
 
                     //Only admins can activate accounts and delete users
-                    if ($this->Auth->isGroup('admin')) {
+                    if ($this->Identity->isGroup('admin')) {
                         //If the user is not active (pending)
                         if (!$user->get('active')) {
                             $actions[] = $this->Form->postLink(__d('me_cms', 'Activate'), ['action' => 'activate', $user->get('id')], [