+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2.31 branch
 ### 2.31.0
 * the `AuthComponent` is deprecated. Use instead the `AuthenticationComponent`;
-* the `AuthHelper` is deprecated.  Use instead the `IdentityHelper`.
+* the `AuthHelper` is deprecated.  Use instead `\MeCms\View\Helper\IdentityHelper`.
 
 ## 2.30 branch
 ### 2.30.13